+++ Rdmp.UI/ProjectUI/Datasets/ConfigureDatasetUI.cs
@@ -493,9 +493,16 @@ namespace Rdmp.UI.ProjectUI.Datasets
 
         private void btnSelectCore_Click(object sender, EventArgs e)
         {
+            var extractionIsFor = SelectedDataSet.ExtractableDataSet.Catalogue_ID;
+
             olvAvailable.SelectObjects(
                 olvAvailable.Objects.OfType<ExtractionInformation>()
-                .Where(ei => ei.ExtractionCategory == ExtractionCategory.Core).ToArray());
+                .Where(ei =>      
+                    //select core columns
+                    ei.ExtractionCategory == ExtractionCategory.Core 
+                    // or ProjectSpecific ones if it is the main dataset
+                    || (extractionIsFor == ei.CatalogueItem.Catalogue_ID && ei.ExtractionCategory == ExtractionCategory.ProjectSpecific)
+                             ).ToArray());
         }
         #region Joins
 
@@ -614,7 +621,7 @@ namespace Rdmp.UI.ProjectUI.Datasets
 
                     ColumnInfo toEmphasise = null;
 
-                    //if theres only one column involved in the join
+                    //if there's only one column involved in the join
                     if (cols.Length == 1)
                         toEmphasise = cols[0]; //emphasise it to the user
                     else
@@ -698,6 +705,9 @@ namespace Rdmp.UI.ProjectUI.Datasets
 
         public void RefreshBus_RefreshObject(object sender, RefreshObjectEventArgs e)
         {
+            if (!SelectedDataSet.Exists())
+                return;
+
             //if an ExtractionInformation is being refreshed
             var ei = e.Object as ExtractionInformation;
             if (ei != null)